refactor: consolidate mobile server architecture and simplify auth
- Remove redundant root files (auth.ts, server.ts) in favor of organized subfolder structure
- Simplify authentication to cookie-only sessions, removing x-session-token header redundancy
- Update all imports to use centralized exports from mobile/index.ts
- Replace monolithic script.js with modular ES6 structure in js/ directory
- Fix clipboard copy path functionality with proper availability detection
- Clean up backup files, test files, and non-relevant documentation
- Maintain same functionality with cleaner, more maintainable codebase

Architecture improvements:
• Centralized mobile server exports through index.ts
• Organized auth management in auth/ subfolder with instance-based patterns
• Modular client-side code: core/, git/, explorer/, utils/ modules
• Better separation of concerns and reduced code duplication

Bug fixes:
• Hide copy path button when clipboard API unavailable (HTTP/insecure contexts)
• Improved error handling for clipboard operations
• Proper clipboard availability detection using navigator.clipboard + isSecureContext
