+++ src/services/mobile/routes/static-routes.ts
@@ -25,7 +25,7 @@ export class StaticRoutes {
             }
             
             if (this.config.useExternalServer && this.config.webPassword) {
-                const sessionToken = req.headers['x-session-token'] as string || (req as any).cookies?.sessionToken;
+                const sessionToken = (req as any).cookies?.sessionToken;
                 if (!sessionToken || !this.authManager.hasActiveSession(sessionToken)) {
                     return res.redirect(`/password?token=${this.config.authToken}`);
                 }
@@ -35,7 +35,7 @@ export class StaticRoutes {
             let html = fs.readFileSync(htmlPath, 'utf8');
             
             html = html.replace('href="styles.css"', `href="styles.css?token=${this.config.authToken}"`);
-            html = html.replace('src="script.js"', `src="script.js?token=${this.config.authToken}"`);
+            html = html.replace('src="js/app.js"', `src="js/app.js?token=${this.config.authToken}"`);
             html = html.replace('href="manifest.json"', `href="manifest.json?token=${this.config.authToken}"`);
             
             html = html.replace('</head>', `
@@ -72,19 +72,6 @@ export class StaticRoutes {
             res.sendFile(path.join(__dirname, '../../../webview/mobile/manifest.json'));
         });
 
-        app.get('/sw.js', (req: Request, res: Response) => {
-            const token = req.query.token || req.headers['x-auth-token'];
-            if (token !== this.config.authToken) {
-                return res.status(401).json({ error: 'Unauthorized' });
-            }
-            
-            if (!this.authManager.checkPasswordForStaticFiles(req, res)) {
-                return;
-            }
-            
-            res.setHeader('Content-Type', 'application/javascript');
-            res.sendFile(path.join(__dirname, '../../../webview/mobile/sw.js'));
-        });
 
         app.get('/styles.css', (req: Request, res: Response) => {
             const token = req.query.token || req.headers['x-auth-token'];
@@ -114,9 +101,25 @@ export class StaticRoutes {
             }
         });
 
-        app.get('/script.js', (req: Request, res: Response) => {
+
+        // Serve modular JavaScript files from /js/ directory
+        app.get(/\/js\/.*/, (req: Request, res: Response) => {
             const token = req.query.token || req.headers['x-auth-token'];
-            if (token !== this.config.authToken) {
+            const referer = (req.headers.referer || req.headers.referrer) as string;
+            
+            // Allow ES6 module imports if:
+            // 1. They have a direct token, OR
+            // 2. They come from an authenticated page (Referer with token), OR  
+            // 3. They come from the same origin (for module import chains)
+            const hasDirectToken = token === this.config.authToken;
+            const fromAuthenticatedPage = referer && referer.includes(`token=${this.config.authToken}`);
+            const fromSameOrigin = referer && (
+                referer.startsWith(`http://${req.headers.host}`) || 
+                referer.startsWith(`https://${req.headers.host}`) ||
+                (req.headers.host && referer.includes(req.headers.host as string))
+            );
+            
+            if (!hasDirectToken && !fromAuthenticatedPage && !fromSameOrigin) {
                 return res.status(401).json({ error: 'Unauthorized' });
             }
             
@@ -124,11 +127,20 @@ export class StaticRoutes {
                 return;
             }
             
-            const filePath = path.join(__dirname, '../../../webview/mobile/script.js');
+            // Extract the file path from the URL
+            const requestedPath = req.path; // e.g., "/js/app.js" or "/js/core/MobileInterface.js"
+            const filePath = path.join(__dirname, '../../../webview/mobile', requestedPath);
+            
+            // Security check: ensure the path stays within the webview/mobile directory
+            const normalizedPath = path.normalize(filePath);
+            const basePath = path.join(__dirname, '../../../webview/mobile');
+            if (!normalizedPath.startsWith(basePath)) {
+                return res.status(403).send('Forbidden: Path traversal detected');
+            }
             
             if (!fs.existsSync(filePath)) {
-                console.error('script.js not found at expected path:', filePath);
-                return res.status(404).send('script.js not found');
+                console.error(`JS file not found at path: ${filePath}`);
+                return res.status(404).send(`JS file not found: ${requestedPath}`);
             }
             
             res.setHeader('Content-Type', 'application/javascript');
@@ -137,8 +149,8 @@ export class StaticRoutes {
                 const jsContent = fs.readFileSync(filePath, 'utf8');
                 res.send(jsContent);
             } catch (error) {
-                console.error('Error reading script.js:', error);
-                res.status(500).send('Error loading script');
+                console.error(`Error reading JS file ${filePath}:`, error);
+                res.status(500).send('Error loading JavaScript file');
             }
         });
     }
